refactor: HttpRequestMessage 클래스 개선
해당 클래스를 message 패키지 아래로 이동시키고, 이름을 `HttpRequest`로 변경함.
